requirements.txt
asgiref==3.4.1
Django==4.0
djangorestframework==3.13.1
joblib==1.1.0
numpy==1.21.5
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.0.0
tzdata==2021.5
xgboost==1.5.1
